Refactor: reorganised files into constants folder
