+++ lib/constants/job_attributes.dart
@@ -1,0 +1,48 @@
+// k prefix is a common naming convention that stands for "constant" in Dart
+
+const List<String> kIndustries = [
+  'Technology',
+  'Healthcare',
+  'Finance',
+  'Education',
+  'Manufacturing',
+  'Retail',
+  'Consulting',
+  'Marketing',
+  'Design',
+  'Sales',
+  'Engineering',
+  'Hospitality',
+  'Legal',
+  'Media',
+  'Other',
+];
+
+const List<String> kWorkArrangements = ['On-site', 'Remote', 'Hybrid'];
+
+const List<String> kJobTypes = [
+  'Full-time',
+  'Part-time',
+  'Internship',
+  'Contract',
+  'Freelance',
+  'Temporary',
+];
+
+const List<String> kDurations = [
+  '1 month',
+  '3 months',
+  '6 months',
+  '1 year',
+  '2+ years',
+  'Permanent',
+];
+
+const List<String> kJobRecencyOptions = [
+  'Any time',
+  'Last 24 hours',
+  'Last 3 days',
+  'Last week',
+  'Last 2 weeks',
+  'Last month',
+];
