+++ lib/screens/auth/forgot_password_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:pronto/constants.dart';
+import 'package:pronto/constants/colours.dart';
 import 'package:pronto/widgets/custom_text_field.dart';
 import 'package:pronto/router.dart';
