+++ lib/screens/profile_setup/applicant/designation_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:pronto/constants.dart';
+import 'package:pronto/constants/colours.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:pronto/widgets/custom_dropdown_field.dart';
 import 'package:pronto/widgets/progress_indicator.dart';
