+++ lib/screens/profile_setup/applicant/social_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:pronto/widgets/progress_indicator.dart';
-import 'package:pronto/constants.dart';
+import 'package:pronto/constants/colours.dart';
 import 'package:pronto/widgets/custom_text_field.dart';
 import 'package:pronto/router.dart';
