+++ lib/widgets/progress_indicator.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:pronto/constants.dart';
+import 'package:pronto/constants/colours.dart';
 
 class CustomProgressIndicator extends StatelessWidget {
   final int currentStep;
